SKSO, syscon cmd exec, jig msg push, silent mode

+++ source/include/clib.h
@@ -3,7 +3,7 @@
 
 #include "types.h"
 
-void* memset(void* s, uint8_t c, uint32_t n);
+void* memset(void* s, uint32_t c, uint32_t n);
 void* memcpy(void* dest, const void* src, uint32_t n);
 int memcmp(const void* s1, const void* s2, uint32_t n);
 uint32_t strlen(const char* str);
